[NO-ISSUE] 계좌 이체 리펙터링

+++ data/home/src/main/java/com/ssafy/home/data/HomeRepositoryImpl.kt
@@ -31,7 +31,7 @@ class HomeRepositoryImpl @Inject constructor(
     private fun HomeResponse.Account.toAccount(): Home.Account {
         return Home.Account(
             accountNo = accountNo,
-            balance = balance,
+            balance = balance.toDouble(),
             currency = currency,
         )
     }
